Add files via upload
Text categorization is the task of classifying a set of documents into categories from a set of predefined labels. Texts cannot be directly handled by our model. The indexing procedure is the first step that maps a text dj into a numeric representation during the training and validation. The standard  TFIDF function is used to represent the text. The unique words from English vocabulary are represented as a dimension of the dataset.